cleanup old db code
