+++ cogs/hunt.py
@@ -16,7 +16,7 @@ import io
 import math
 import time
 
-from utils.db2 import DBase
+from utils.db import DBase
 from utils.drive import Drive
 import os
